feat: lazily load portfolio sections for faster TTI
