+++ src/App.tsx
@@ -1,7 +1,9 @@
+import React from "react";
 import CommandK, { type CommandKItem } from "@/components/command/CommandK";
-import Portfolio from "./Portfolio";
 import { AnimatePresence, motion } from "framer-motion";
 import { useLocation } from "react-router-dom";
+
+const Portfolio = React.lazy(() => import("./Portfolio"));
 
 const COMMAND_ITEMS: CommandKItem[] = [
   {
@@ -41,9 +43,23 @@ export default function App() {
           exit={{ opacity: 0, y: -8 }}
           transition={{ duration: 0.2, ease: "easeOut" }}
         >
-          <Portfolio />
+          <React.Suspense fallback={<PageFallback />}>
+            <Portfolio />
+          </React.Suspense>
         </motion.main>
       </AnimatePresence>
     </>
+  );
+}
+
+function PageFallback() {
+  return (
+    <div className="mx-auto flex min-h-[60vh] max-w-6xl items-center justify-center px-4">
+      <div
+        className="h-12 w-12 animate-spin rounded-full border-4 border-primary/30 border-t-primary"
+        role="status"
+        aria-label="초기 화면 로딩 중"
+      />
+    </div>
   );
 }
